mail ready for use

+++ xcpvmbackup.sh
@@ -1,43 +1,49 @@
 #!/bin/bash
 #
-# Written By: Mr Rahul Kumar
-# Created date: Jun 14, 2014
-# Last Updated: Mar 08, 2017
-# Version: 1.2.1
-# Visit: https://tecadmin.net/backup-running-virtual-machine-in-xenserver/
-#
 # Fork: Boardereb
 # Only tested on XCP-NG
+# GPLv2+
+# V: 0.01-Beta1
+##############################
+# Pre-Fork:
+# Written By: Mr Rahul Kumar
+# Created date: Jun 14, 2014
+# Visit: https://tecadmin.net/backup-running-virtual-machine-in-xenserver/
+##############################
 
-# To change:
+
+### To change:
 
 NFS_SERVER_IP="192.168.10.100"
-MOUNTPOINT=/mnt/nfs
+MOUNTPOINT="/mnt/nfs"
 FILE_LOCATION_ON_NFS="/remote/nfs/location"
-MAXBACKUPS=2
+MAXBACKUPS="2"
 
-# Loglevel 0=Only Errors
-# Loglevel 1=start and exit and Errors
-# Loglevel 2=every action
+### Loglevel 0=Only Errors
+### Loglevel 1=start and exit and Errors
+### Loglevel 2=every action
 
-LOGLEVEL=1
+LOGLEVEL="1"
 SYSLOGER="true"
-LOGMAIL="true"
-#MAILADRESS="your@email.com" # if not set send mail to $USER
 
-#SET SYSLOGGERSYSLOGGERPATH IF NOT "logger"
+### Send Log-Email with mailx - make sure it is configured:
+### echo "bash testmail" | mail -s "testmail from bash" "your@email.com"
+#LOGMAIL="true"
+#MAILADRESS="your@email.com" ### if not set send mail to $USER
+
+####Set SysLoggerPath if not "logger"
 #SYSLOGGERPATH="/usr/bin/logger"
 
-# int. Variablen
+### int. Variablen
 
 XSNAME=`echo $HOSTNAME`
 DATE=$(date +%d-%m-%Y)-$(cat /dev/urandom | tr -dc 'a-zA-Z0-9' | fold -w 10 | head -n 1)
 UUIDFILE=$(mktemp /tmp/uuids.XXXXXXXXX)
 MAILFILE=$(mktemp /tmp/mail.XXXXXXXXX)
 
-## LOGGERMASSAGE function
-## Var1: LOGGERMASSAGE LOGLEVEL "LOGMASSAGE"
-## Var2: LOGGERMASSAGE "LOGMASSAGE" = LOGLEVEL 1 = only start and stop
+### LOGGERMASSAGE function
+### Var1: LOGGERMASSAGE LOGLEVEL "LOGMASSAGE"
+### Var2: LOGGERMASSAGE "LOGMASSAGE" = LOGLEVEL 1 = only start and stop
 
 function SYSLOGGER() {
   if [[ SYSLOGER -eq "true" ]]; then
@@ -54,23 +60,19 @@ function SYSLOGGER() {
 }
 
 function LOGGERMASSAGE() {
-  if [[ $LOGLEVEL > 1 ]]; then
-    if [[ $1 =~ ^[0-9]+$ ]]; then
-    	echo $0: $2
-      echo "$0: $2" >> $MAILFILE
-	 	  SYSLOGGER "$0: $2"
-    else
-      echo $0: $1 $2
-      echo "$0: $1 $2" >> $MAILFILE
-	 	  SYSLOGGER "$0: $1 $2"
-    fi
+  local number='^[0-9]+$'
+  if [[ $1 =~ $number ]]; then
+    local LOGGERMASSAGELEVEL="$1"
+    local LOGGERMASSAGE="$2"
   else
-    if [[ $1 -le $LOGLEVEL ]]; then
-      echo $0: $2
-      echo "$0: $2" >> $MAILFILE
-      SYSLOGGER "$0: $2"
-    fi
+    local LOGGERMASSAGELEVEL="2"
+    local LOGGERMASSAGE="$1"
   fi
+  if [[ $LOGGERMASSAGELEVEL -le $LOGLEVEL ]]; then
+    echo $0: $LOGGERMASSAGE
+    echo $0: $LOGGERMASSAGE >> $MAILFILE
+    SYSLOGGER "$0: $LOGGERMASSAGE"
+    fi
 }
 
 ### Send Mail
@@ -125,9 +127,7 @@ function FREESPACE() {
   for DISK in "$DISKLIST"; do
     DISKSPACE=$(echo $DISK $DISKSPACE | awk '{print $1 + $2}')
   done
-
   FREESPACE=$(df --block-size=1 $2 --output=avail | sed -e 1d)
-
   if [[ $DISKSPACE -lt  $FREESPACE ]]; then
     echo "true"
   else
@@ -147,24 +147,23 @@ fi
 
 ### check if nfs allrady moundet if not mount
 MOUNDET=$(stat -c%d "$MOUNTPOINT")
-if grep -qs "$NFS_SERVER_IP:$FILE_LOCATION_ON_NFS $MOUNTPOINT" /proc/mounts; then
+MOUNT=$(grep "$NFS_SERVER_IP:$FILE_LOCATION_ON_NFS $MOUNTPOINT" /proc/mounts | grep $MOUNTPOINT)
+if [[ -z $MOUNT ]]; then
+    LOGGERMASSAGE "mount NFS $NFS_SERVER_IP:$FILE_LOCATION_ON_NFS"
+    mount -t nfs $NFS_SERVER_IP:$FILE_LOCATION_ON_NFS $MOUNTPOINT
+    if [[ `stat -c%d "$MOUNTPOINT"` -eq $MOUNDET ]]; then
+      LOGGERMASSAGE 0 "Error: Coult not mount $NFS_SERVER_IP:$FILE_LOCATION_ON_NFS $MOUNTPOINT"
+      QUIT 1
+    fi
+else
   MOUNDET="allrady"
   LOGGERMASSAGE "$NFS_SERVER_IP:$FILE_LOCATION_ON_NFS $MOUNTPOINT allrady mounted"
-else
-  LOGGERMASSAGE "mount NFS $NFS_SERVER_IP:$FILE_LOCATION_ON_NFS"
-  mount -t nfs $NFS_SERVER_IP:$FILE_LOCATION_ON_NFS $MOUNTPOINT
-fi
-
-if [[ `stat -c%d "$MOUNTPOINT"` -eq $MOUNDET ]]; then
-  LOGGERMASSAGE 0 "Error: Coult not mount $NFS_SERVER_IP:$FILE_LOCATION_ON_NFS $MOUNTPOINT"
-  QUIT 1
 fi
 
 ### creat backuppath if not exist
 BACKUPPATH="$MOUNTPOINT/$XSNAME/$DATE"
 LOGGERMASSAGE "create backuppath $BACKUPPATH if not exist"
 mkdir -p $BACKUPPATH
-
 if [[ ! -d ${BACKUPPATH} ]]; then
 	LOGGERMASSAGE 0 "Error: No backup directory found"
 	QUIT 1
@@ -185,41 +184,58 @@ do
   if [[ $(FREESPACE $VMUUID /mnt/nfsbackup) == "true" ]]; then
     LOGGERMASSAGE "create snapshoot from: $VMNAME"
     SNAPUUID=`xe vm-snapshot uuid=$VMUUID new-name-label="SNAPSHOT-$VMNAME-$DATE"`
-  	xe template-param-set is-a-template=false ha-always-run=false uuid=${SNAPUUID}
+    xe template-param-set is-a-template=false ha-always-run=false uuid=${SNAPUUID}
+    if [[ $? -ne 0 ]]; then
+      EXPORTERROR="true"
+    fi
 
   	LOGGERMASSAGE "export snapshoot $VMNAME to $BACKUPPATH"
   	xe vm-export vm=${SNAPUUID} filename="$BACKUPPATH/$VMNAME-$DATE.xva"
+    if [[ $? -ne 0 ]]; then
+      EXPORTERROR="true"
+    fi
 
-  	LOGGERMASSAGE "remove snapshoot from: $VMNAME"
+    LOGGERMASSAGE "remove snapshoot from: $VMNAME"
   	xe vm-uninstall uuid=${SNAPUUID} force=true
+    if [[ $? -ne 0 ]]; then
+      EXPORTERROR="true"
+    fi
   else
+    EXPORTERROR="true"
     LOGGERMASSAGE 0 "Error: not enough space to export $VMNAME to $MOUNTPOINT"
   fi
 done < ${UUIDFILE}
 
 ## start remove old backups
-BACKUPS=$(BACKUPDIRS -c)
-COUNT=$(($BACKUPS-$MAXBACKUPS))
-if [[ $COUNT -lt 0 ]]; then
-  COUNT=0
-fi
-LOGGERMASSAGE "$BACKUPS backup found - remove $COUNT old backup"
 
-if [[ $COUNT > 0 ]]; then
-  COUNT=1
-  for DIR in $(BACKUPDIRS);do
-    if [[ $COUNT > $MAXBACKUPS ]]; then
-      rm -rf $DIR
-    fi
-    COUNT=$(($COUNT+1))
-  done
-
-  if [[ $(BACKUPDIRS -c) == $MAXBACKUPS ]]; then
-    LOGGERMASSAGE "old backups are removed"
-  else
-    LOGGERMASSAGE 0 "Error: not all old backups are removed"
+if [[ -z "$EXPORTERROR" ]]; then
+  BACKUPS=$(BACKUPDIRS -c)
+  COUNT=$(($BACKUPS-$MAXBACKUPS))
+  if [[ $COUNT -lt 0 ]]; then
+    COUNT=0
   fi
+  LOGGERMASSAGE "$BACKUPS backup found - remove $COUNT old backup"
+
+  if [[ $COUNT > 0 ]]; then
+    COUNT=1
+    for DIR in $(BACKUPDIRS);do
+      if [[ $COUNT > $MAXBACKUPS ]]; then
+        rm -rf $DIR
+      fi
+      COUNT=$(($COUNT+1))
+    done
+
+    if [[ $(BACKUPDIRS -c) == $MAXBACKUPS ]]; then
+      LOGGERMASSAGE "old backups are removed"
+    else
+      LOGGERMASSAGE 0 "Error: not all old backups are removed"
+    fi
+  fi
+else
+      LOGGERMASSAGE 0 "Error: Do not remove old Backups becouse a error in the new backup"
 fi
+
+
 
 # unmount if not befor for the script moundet
 if [[ $MOUNDET != "allrady"  ]]; then
